requirements.txt
pandas==2.0.3
azure-storage-blob==12.17.0
python-dotenv==1.0.0
google-generativeai==0.7.2
streamlit==1.28.1
openpyxl==3.1.2
pyarrow==12.0.1
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.0
scikit-learn==1.3.0
numpy==1.24.3
tabulate==0.9.0